+++ src/common/printing.c
@@ -100,7 +100,7 @@ void ffPrintFormat(const char* moduleName, uint8_t moduleIndex, const FFModuleAr
     ffStrbufPutTo(&buffer, stdout);
 }
 
-static void printError(const char* moduleName, uint8_t moduleIndex, const FFModuleArgs* moduleArgs, FFPrintType printType, const char* message, va_list arguments)
+void ffPrintError(const char* moduleName, uint8_t moduleIndex, const FFModuleArgs* moduleArgs, FFPrintType printType, const char* message, ...)
 {
     if(!instance.config.display.showErrors)
         return;
@@ -110,20 +110,15 @@ static void printError(const char* moduleName, uint8_t moduleIndex, const FFModu
     if(!instance.config.display.pipe)
         fputs(FASTFETCH_TEXT_MODIFIER_ERROR, stdout);
 
+    va_list arguments;
+    va_start(arguments, message);
     vprintf(message, arguments);
+    va_end(arguments);
 
     if(!instance.config.display.pipe)
         fputs(FASTFETCH_TEXT_MODIFIER_RESET, stdout);
 
     putchar('\n');
-}
-
-void ffPrintError(const char* moduleName, uint8_t moduleIndex, const FFModuleArgs* moduleArgs, FFPrintType printType, const char* message, ...)
-{
-    va_list arguments;
-    va_start(arguments, message);
-    printError(moduleName, moduleIndex, moduleArgs, printType, message, arguments);
-    va_end(arguments);
 }
 
 void ffPrintColor(const FFstrbuf* colorValue)