+++ src/detection/wifi/wifi_nbsd.c
@@ -2,12 +2,21 @@
 #include "common/io/io.h"
 #include "util/stringUtils.h"
 
+#define COMPAT_FREEBSD_NET80211 1
 #include <sys/ioctl.h>
 #include <sys/socket.h>
 #include <net/if.h>
+#include <net/if_media.h>
 #include <net80211/ieee80211.h>
 #include <net80211/ieee80211_ioctl.h>
-#include <unistd.h>
+
+// ieee80211 header of NetBSD is full of mess. Add compatibility macros from FreeBSD
+#undef IEEE80211_IS_CHAN_ANYG
+#define IEEE80211_IS_CHAN_ANYG(x) (IEEE80211_IS_CHAN_PUREG(x) || IEEE80211_IS_CHAN_G(x))
+#undef IEEE80211_IS_CHAN_HT
+#define IEEE80211_IS_CHAN_HT(x) (((x)->ic_flags & IEEE80211_CHAN_HT) != 0)
+#undef IEEE80211_IS_CHAN_VHT
+#define IEEE80211_IS_CHAN_VHT(x) (((x)->ic_flags & IEEE80211_CHAN_VHT) != 0)
 
 const char* ffDetectWifi(FFlist* result)
 {
@@ -18,6 +27,7 @@ const char* ffDetectWifi(FFlist* result)
 
     FF_AUTO_CLOSE_FD int sock = socket(AF_INET, SOCK_DGRAM, 0);
     if(sock < 0) {
+        if_freenameindex(infs);
         return "socket() failed";
     }
 
@@ -41,67 +51,115 @@ const char* ffDetectWifi(FFlist* result)
         item->conn.channel = 0;
         item->conn.frequency = 0;
 
-        struct ieee80211_nodereq nr = {};
-        strlcpy(nr.nr_ifname, i->if_name, sizeof(nr.nr_ifname));
+        char ssid[IEEE80211_NWID_LEN + 1] = {};
+        struct ieee80211req ireq = {};
+        strlcpy(ireq.i_name, i->if_name, sizeof(ireq.i_name));
+        ireq.i_type = IEEE80211_IOC_SSID;
+        ireq.i_data = ssid;
+        ireq.i_len = sizeof(ssid) - 1;
 
-        struct ifreq ifr = {};
-        strlcpy(ifr.ifr_name, i->if_name, sizeof(ifr.ifr_name));
-        if (ioctl(sock, SIOCGIFFLAGS, &ifr) < 0) {
-            ffStrbufSetStatic(&item->inf.status, "Unknown");
-        } else {
-            ffStrbufSetStatic(&item->inf.status, ifr.ifr_flags & IFF_UP ? "Up" : "Down");
-        }
-
-        if (ioctl(sock, SIOCG80211NODE, &nr) < 0) {
-            ffStrbufSetStatic(&item->conn.status, "Not associated");
+        if (ioctl(sock, SIOCG80211, &ireq) < 0 || ireq.i_len == 0) {
+            struct ifreq ifr;
+            strlcpy(ifr.ifr_name, i->if_name, sizeof(ifr.ifr_name));
+            if (ioctl(sock, SIOCGIFFLAGS, &ifr) < 0) {
+                ffStrbufSetStatic(&item->inf.status, "Unknown");
+            } else {
+                ffStrbufSetStatic(&item->inf.status, ifr.ifr_flags & IFF_UP ? "Up" : "Down");
+            }
+            ffStrbufAppendS(&item->conn.status, "Not associated");
             continue;
         }
 
-        if (nr.nr_nwid_len > 0) {
-            ffStrbufSetStatic(&item->conn.status, "Associated");
-            ffStrbufAppendNS(&item->conn.ssid, nr.nr_nwid_len, (char*)nr.nr_nwid);
-        } else {
-            ffStrbufSetStatic(&item->conn.status, "Not associated");
-            continue;
+        ffStrbufSetStatic(&item->inf.status, "Up");
+        ffStrbufSetStatic(&item->conn.status, "Associated");
+        ffStrbufAppendNS(&item->conn.ssid, ireq.i_len, ssid);
+
+        uint8_t bssid[IEEE80211_ADDR_LEN] = {};
+        ireq.i_type = IEEE80211_IOC_BSSID;
+        ireq.i_data = bssid;
+        ireq.i_len = sizeof(bssid);
+
+        if (ioctl(sock, SIOCG80211, &ireq) >= 0) {
+            ffStrbufSetF(&item->conn.bssid, "%02X:%02X:%02X:%02X:%02X:%02X",
+                         bssid[0], bssid[1], bssid[2], bssid[3], bssid[4], bssid[5]);
         }
 
-        ffStrbufSetF(&item->conn.bssid, "%02X:%02X:%02X:%02X:%02X:%02X",
-                    nr.nr_bssid[0], nr.nr_bssid[1], nr.nr_bssid[2],
-                    nr.nr_bssid[3], nr.nr_bssid[4], nr.nr_bssid[5]);
+        struct ieee80211_channel curchan = {};
+        ireq.i_type = IEEE80211_IOC_CHANNEL;
+        ireq.i_data = &curchan;
+        ireq.i_len = sizeof(curchan);
 
-        item->conn.channel = nr.nr_channel;
+        if (ioctl(sock, SIOCG80211, &ireq) >= 0) {
+            // item->conn.channel = curchan.ic_ieee; // No ic_ieee in NetBSD
+            item->conn.channel = ffWifiFreqToChannel(curchan.ic_freq);
+            item->conn.frequency = curchan.ic_freq;
 
-        if (nr.nr_max_rssi) {
-            item->conn.signalQuality = ((float)nr.nr_rssi / nr.nr_max_rssi) * 100.0;
+            #ifdef IEEE80211_IS_CHAN_HE // for future use
+            if (IEEE80211_IS_CHAN_HE(&curchan))
+                ffStrbufSetStatic(&item->conn.protocol, "802.11ax (Wi-Fi 6)");
+            else
+            #endif
+            if (IEEE80211_IS_CHAN_VHT(&curchan))
+                ffStrbufSetStatic(&item->conn.protocol, "802.11ac (Wi-Fi 5)");
+            else if (IEEE80211_IS_CHAN_HT(&curchan))
+                ffStrbufSetStatic(&item->conn.protocol, "802.11n (Wi-Fi 4)");
+            else if (IEEE80211_IS_CHAN_ANYG(&curchan))
+                ffStrbufSetStatic(&item->conn.protocol, "802.11g");
+            else if (IEEE80211_IS_CHAN_B(&curchan))
+                ffStrbufSetStatic(&item->conn.protocol, "802.11b");
+            else if (IEEE80211_IS_CHAN_A(&curchan))
+                ffStrbufSetStatic(&item->conn.protocol, "802.11a");
+            else if (IEEE80211_IS_CHAN_FHSS(&curchan))
+                ffStrbufSetStatic(&item->conn.protocol, "802.11 (FHSS)");
         }
 
-        if (nr.nr_flags & IEEE80211_NODEREQ_HT) {
-            ffStrbufSetStatic(&item->conn.protocol, "802.11n (Wi-Fi 4)");
-        } else if (nr.nr_flags & IEEE80211_NODEREQ_VHT) {
-            ffStrbufSetStatic(&item->conn.protocol, "802.11ac (Wi-Fi 5)");
-        } else if (nr.nr_chan_flags & IEEE80211_CHANINFO_5GHZ) {
-            ffStrbufSetStatic(&item->conn.protocol, "802.11a");
-        } else if (nr.nr_chan_flags & IEEE80211_CHANINFO_2GHZ) {
-            ffStrbufSetStatic(&item->conn.protocol, "802.11g");
+        union {
+            struct ieee80211req_sta_req req;
+            uint8_t buf[1024];
+        } stareq = {};
+        memcpy(stareq.req.is_u.macaddr, bssid, sizeof(bssid));
+        ireq.i_type = IEEE80211_IOC_STA_INFO;
+        ireq.i_data = &stareq;
+        ireq.i_len = sizeof(stareq);
+
+        if (ioctl(sock, SIOCG80211, &ireq) >= 0) {
+            struct ieee80211req_sta_info* sta = stareq.req.info;
+            if (sta->isi_len != 0) {
+                int8_t rssi = (int8_t) sta->isi_rssi; // Strangely, `sta->isi_rssi` is unsigned
+                item->conn.signalQuality = (rssi >= -50 ? 100 : rssi <= -100 ? 0 : (rssi + 100) * 2);
+
+                if (sta->isi_txrate) {
+                    item->conn.txRate = (double)sta->isi_txrate / 2.0;
+                }
+            }
         }
 
-        struct ieee80211_wpaparams wpa = {};
-        strlcpy(wpa.i_name, i->if_name, sizeof(wpa.i_name));
-
-        if (ioctl(sock, SIOCG80211WPAPARMS, &wpa) >= 0 && wpa.i_enabled) {
-            if (wpa.i_protos & IEEE80211_WPA_PROTO_WPA2)
-                ffStrbufSetStatic(&item->conn.security, "WPA2");
-            else if (wpa.i_protos & IEEE80211_WPA_PROTO_WPA1)
+        ireq.i_type = IEEE80211_IOC_AUTHMODE;
+        ireq.i_data = NULL;
+        ireq.i_len = 0;
+        if (ioctl(sock, SIOCG80211, &ireq) >= 0) {
+            switch (ireq.i_val) {
+            case IEEE80211_AUTH_NONE:
+                ffStrbufSetStatic(&item->conn.security, "Insecure");
+                break;
+            case IEEE80211_AUTH_OPEN:
+                ffStrbufSetStatic(&item->conn.security, "Open");
+                break;
+            case IEEE80211_AUTH_SHARED:
+                ffStrbufSetStatic(&item->conn.security, "Shared");
+                break;
+            case IEEE80211_AUTH_8021X:
+                ffStrbufSetStatic(&item->conn.security, "8021X");
+                break;
+            case IEEE80211_AUTH_AUTO:
+                ffStrbufSetStatic(&item->conn.security, "Auto");
+                break;
+            case IEEE80211_AUTH_WPA:
                 ffStrbufSetStatic(&item->conn.security, "WPA");
-        } else {
-            struct ieee80211_nwkey nwkey = {};
-            strlcpy(nwkey.i_name, i->if_name, sizeof(nwkey.i_name));
-
-            if (ioctl(sock, SIOCG80211NWKEY, &nwkey) >= 0) {
-                if (nwkey.i_wepon)
-                    ffStrbufSetStatic(&item->conn.security, "WEP");
-                else
-                    ffStrbufSetStatic(&item->conn.security, "Open");
+                break;
+            default:
+                ffStrbufSetF(&item->conn.security, "Unknown (%d)", ireq.i_val);
+                break;
             }
         }
     }